@@ -44,6 +44,7 @@ import com.github._1c_syntax.bsl.languageserver.providers.ReferencesProvider;
 import com.github._1c_syntax.bsl.languageserver.providers.RenameProvider;
 import com.github._1c_syntax.bsl.languageserver.providers.SelectionRangeProvider;
 import com.github._1c_syntax.bsl.languageserver.utils.Ranges;
+import jakarta.annotation.PreDestroy;
 import lombok.RequiredArgsConstructor;
 import org.eclipse.lsp4j.CallHierarchyIncomingCall;
 import org.eclipse.lsp4j.CallHierarchyIncomingCallsParams;
@@ -93,13 +94,15 @@ import org.eclipse.lsp4j.WorkspaceEdit;
 import org.eclipse.lsp4j.jsonrpc.messages.Either;
 import org.eclipse.lsp4j.jsonrpc.messages.Either3;
 import org.eclipse.lsp4j.services.TextDocumentService;
+import org.springframework.scheduling.concurrent.CustomizableThreadFactory;
 import org.springframework.stereotype.Component;
 
 import java.net.URI;
 import java.util.Collections;
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
-import java.util.stream.Collectors;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
 
 @Component
 @RequiredArgsConstructor
@@ -123,14 +126,22 @@ public class BSLTextDocumentService implements TextDocumentService, ProtocolExte
   private final RenameProvider renameProvider;
   private final InlayHintProvider inlayHintProvider;
 
+  private final ExecutorService executorService = Executors.newCachedThreadPool(new CustomizableThreadFactory("text-document-service-"));
+
+  @PreDestroy
+  private void onDestroy() {
+    executorService.shutdown();
+  }
+
   @Override
   public CompletableFuture<Hover> hover(HoverParams params) {
     var documentContext = context.getDocument(params.getTextDocument().getUri());
     if (documentContext == null) {
       return CompletableFuture.completedFuture(null);
     }
-    return CompletableFuture.supplyAsync(() ->
-      hoverProvider.getHover(documentContext, params).orElse(null)
+    return CompletableFuture.supplyAsync(
+      () -> hoverProvider.getHover(documentContext, params).orElse(null),
+      executorService
     );
   }
 
@@ -143,8 +154,9 @@ public class BSLTextDocumentService implements TextDocumentService, ProtocolExte
       return CompletableFuture.completedFuture(Either.forRight(Collections.emptyList()));
     }
 
-    return CompletableFuture.supplyAsync(() ->
-      Either.forRight(definitionProvider.getDefinition(documentContext, params))
+    return CompletableFuture.supplyAsync(
+      () -> Either.forRight(definitionProvider.getDefinition(documentContext, params)),
+      executorService
     );
   }
 
@@ -155,7 +167,10 @@ public class BSLTextDocumentService implements TextDocumentService, ProtocolExte
       return CompletableFuture.completedFuture(Collections.emptyList());
     }
 
-    return CompletableFuture.supplyAsync(() -> referencesProvider.getReferences(documentContext, params));
+    return CompletableFuture.supplyAsync(
+      () -> referencesProvider.getReferences(documentContext, params),
+      executorService
+    );
   }
 
   @Override
@@ -167,10 +182,11 @@ public class BSLTextDocumentService implements TextDocumentService, ProtocolExte
       return CompletableFuture.completedFuture(null);
     }
 
-    return CompletableFuture.supplyAsync(() ->
-      documentSymbolProvider.getDocumentSymbols(documentContext).stream()
+    return CompletableFuture.supplyAsync(
+      () -> documentSymbolProvider.getDocumentSymbols(documentContext).stream()
         .map(Either::<SymbolInformation, DocumentSymbol>forRight)
-        .collect(Collectors.toList())
+        .toList(),
+      executorService
     );
   }
 
@@ -181,7 +197,10 @@ public class BSLTextDocumentService implements TextDocumentService, ProtocolExte
       return CompletableFuture.completedFuture(null);
     }
 
-    return CompletableFuture.supplyAsync(() -> codeActionProvider.getCodeActions(params, documentContext));
+    return CompletableFuture.supplyAsync(
+      () -> codeActionProvider.getCodeActions(params, documentContext),
+      executorService
+    );
   }
 
   @Override
@@ -191,7 +210,10 @@ public class BSLTextDocumentService implements TextDocumentService, ProtocolExte
       return CompletableFuture.completedFuture(Collections.emptyList());
     }
 
-    return CompletableFuture.supplyAsync(() -> codeLensProvider.getCodeLens(documentContext));
+    return CompletableFuture.supplyAsync(
+      () -> codeLensProvider.getCodeLens(documentContext),
+      executorService
+    );
   }
 
   @Override
@@ -201,7 +223,10 @@ public class BSLTextDocumentService implements TextDocumentService, ProtocolExte
     if (documentContext == null) {
       return CompletableFuture.completedFuture(unresolved);
     }
-    return CompletableFuture.supplyAsync(() -> codeLensProvider.resolveCodeLens(documentContext, unresolved, data));
+    return CompletableFuture.supplyAsync(
+      () -> codeLensProvider.resolveCodeLens(documentContext, unresolved, data),
+      executorService
+    );
   }
 
   @Override
@@ -211,8 +236,10 @@ public class BSLTextDocumentService implements TextDocumentService, ProtocolExte
       return CompletableFuture.completedFuture(null);
     }
 
-    List<TextEdit> edits = formatProvider.getFormatting(params, documentContext);
-    return CompletableFuture.completedFuture(edits);
+    return CompletableFuture.supplyAsync(
+      () -> formatProvider.getFormatting(params, documentContext),
+      executorService
+    );
   }
 
   @Override
@@ -222,8 +249,10 @@ public class BSLTextDocumentService implements TextDocumentService, ProtocolExte
       return CompletableFuture.completedFuture(null);
     }
 
-    List<TextEdit> edits = formatProvider.getRangeFormatting(params, documentContext);
-    return CompletableFuture.completedFuture(edits);
+    return CompletableFuture.supplyAsync(
+      () -> formatProvider.getRangeFormatting(params, documentContext),
+      executorService
+    );
   }
 
   @Override
@@ -233,7 +262,10 @@ public class BSLTextDocumentService implements TextDocumentService, ProtocolExte
       return CompletableFuture.completedFuture(null);
     }
 
-    return CompletableFuture.supplyAsync(() -> foldingRangeProvider.getFoldingRange(documentContext));
+    return CompletableFuture.supplyAsync(
+      () -> foldingRangeProvider.getFoldingRange(documentContext),
+      executorService
+    );
   }
 
   @Override
@@ -244,13 +276,16 @@ public class BSLTextDocumentService implements TextDocumentService, ProtocolExte
       return CompletableFuture.completedFuture(null);
     }
 
-    return CompletableFuture.supplyAsync(() -> {
-      List<CallHierarchyItem> callHierarchyItems = callHierarchyProvider.prepareCallHierarchy(documentContext, params);
-      if (callHierarchyItems.isEmpty()) {
-        return null;
-      }
-      return callHierarchyItems;
-    });
+    return CompletableFuture.supplyAsync(
+      () -> {
+        List<CallHierarchyItem> callHierarchyItems = callHierarchyProvider.prepareCallHierarchy(documentContext, params);
+        if (callHierarchyItems.isEmpty()) {
+          return null;
+        }
+        return callHierarchyItems;
+      },
+      executorService
+    );
   }
 
   @Override
@@ -262,7 +297,10 @@ public class BSLTextDocumentService implements TextDocumentService, ProtocolExte
       return CompletableFuture.completedFuture(Collections.emptyList());
     }
 
-    return CompletableFuture.supplyAsync(() -> callHierarchyProvider.incomingCalls(documentContext, params));
+    return CompletableFuture.supplyAsync(
+      () -> callHierarchyProvider.incomingCalls(documentContext, params),
+      executorService
+    );
   }
 
   @Override
@@ -274,7 +312,10 @@ public class BSLTextDocumentService implements TextDocumentService, ProtocolExte
       return CompletableFuture.completedFuture(Collections.emptyList());
     }
 
-    return CompletableFuture.supplyAsync(() -> callHierarchyProvider.outgoingCalls(documentContext, params));
+    return CompletableFuture.supplyAsync(
+      () -> callHierarchyProvider.outgoingCalls(documentContext, params),
+      executorService
+    );
   }
 
   @Override
@@ -284,7 +325,10 @@ public class BSLTextDocumentService implements TextDocumentService, ProtocolExte
       return CompletableFuture.completedFuture(Collections.emptyList());
     }
 
-    return CompletableFuture.supplyAsync(() -> selectionRangeProvider.getSelectionRange(documentContext, params));
+    return CompletableFuture.supplyAsync(
+      () -> selectionRangeProvider.getSelectionRange(documentContext, params),
+      executorService
+    );
   }
 
   @Override
@@ -294,7 +338,10 @@ public class BSLTextDocumentService implements TextDocumentService, ProtocolExte
       return CompletableFuture.completedFuture(Collections.emptyList());
     }
 
-    return CompletableFuture.supplyAsync(() -> colorProvider.getDocumentColor(documentContext));
+    return CompletableFuture.supplyAsync(
+      () -> colorProvider.getDocumentColor(documentContext),
+      executorService
+    );
   }
 
   @Override
@@ -304,7 +351,10 @@ public class BSLTextDocumentService implements TextDocumentService, ProtocolExte
       return CompletableFuture.completedFuture(Collections.emptyList());
     }
 
-    return CompletableFuture.supplyAsync(() -> colorProvider.getColorPresentation(documentContext, params));
+    return CompletableFuture.supplyAsync(
+      () -> colorProvider.getColorPresentation(documentContext, params),
+      executorService
+    );
   }
 
   @Override
@@ -314,7 +364,10 @@ public class BSLTextDocumentService implements TextDocumentService, ProtocolExte
       return CompletableFuture.completedFuture(Collections.emptyList());
     }
 
-    return CompletableFuture.supplyAsync(() -> inlayHintProvider.getInlayHint(documentContext, params));
+    return CompletableFuture.supplyAsync(
+      () -> inlayHintProvider.getInlayHint(documentContext, params),
+      executorService
+    );
   }
 
   @Override
@@ -380,7 +433,10 @@ public class BSLTextDocumentService implements TextDocumentService, ProtocolExte
       return CompletableFuture.completedFuture(null);
     }
 
-    return CompletableFuture.supplyAsync(() -> documentLinkProvider.getDocumentLinks(documentContext));
+    return CompletableFuture.supplyAsync(
+      () -> documentLinkProvider.getDocumentLinks(documentContext),
+      executorService
+    );
   }
 
   @Override
@@ -397,7 +453,7 @@ public class BSLTextDocumentService implements TextDocumentService, ProtocolExte
       if (range != null) {
         diagnostics = diagnostics.stream()
           .filter(diagnostic -> Ranges.containsRange(range, diagnostic.getRange()))
-          .collect(Collectors.toList());
+          .toList();
       }
       return new Diagnostics(diagnostics, documentContext.getVersion());
     });
@@ -410,8 +466,10 @@ public class BSLTextDocumentService implements TextDocumentService, ProtocolExte
       return CompletableFuture.completedFuture(null);
     }
 
-    return CompletableFuture.supplyAsync(() ->
-      Either3.forFirst(renameProvider.getPrepareRename(documentContext, params)));
+    return CompletableFuture.supplyAsync(
+      () -> Either3.forFirst(renameProvider.getPrepareRename(documentContext, params)),
+      executorService
+    );
   }
 
   @Override
@@ -421,7 +479,10 @@ public class BSLTextDocumentService implements TextDocumentService, ProtocolExte
       return CompletableFuture.completedFuture(null);
     }
 
-    return CompletableFuture.supplyAsync(() -> renameProvider.getRename(documentContext, params));
+    return CompletableFuture.supplyAsync(
+      () -> renameProvider.getRename(documentContext, params),
+      executorService
+    );
   }
 
   public void reset() {
