@@ -21,12 +21,12 @@
  */
 package com.github._1c_syntax.bsl.languageserver;
 
+import jakarta.annotation.PostConstruct;
 import lombok.EqualsAndHashCode;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
-import javax.annotation.PostConstruct;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.Serializable;
