@@ -21,8 +21,9 @@
  */
 package com.github._1c_syntax.bsl.languageserver.websocket;
 
+import jakarta.websocket.server.ServerEndpointConfig;
 import org.eclipse.lsp4j.services.LanguageClient;
-import org.eclipse.lsp4j.websocket.WebSocketEndpoint;
+import org.eclipse.lsp4j.websocket.jakarta.WebSocketEndpoint;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnWebApplication;
 import org.springframework.context.annotation.Bean;
@@ -30,8 +31,6 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.socket.config.annotation.EnableWebSocket;
 import org.springframework.web.socket.server.standard.ServerEndpointExporter;
 import org.springframework.web.socket.server.standard.ServerEndpointRegistration;
-
-import javax.websocket.server.ServerEndpointConfig;
 
 /**
  * Конфигурация модуля веб-сокетов.
