@@ -22,6 +22,13 @@
 package com.github._1c_syntax.bsl.languageserver.websocket;
 
 import com.github._1c_syntax.bsl.languageserver.util.CleanupContextBeforeClassAndAfterEachTestMethod;
+import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
+import jakarta.websocket.ClientEndpoint;
+import jakarta.websocket.ContainerProvider;
+import jakarta.websocket.DeploymentException;
+import jakarta.websocket.OnOpen;
+import jakarta.websocket.Session;
+import jakarta.websocket.WebSocketContainer;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -30,12 +37,6 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.test.context.ActiveProfiles;
 
-import javax.websocket.ClientEndpoint;
-import javax.websocket.ContainerProvider;
-import javax.websocket.DeploymentException;
-import javax.websocket.OnOpen;
-import javax.websocket.Session;
-import javax.websocket.WebSocketContainer;
 import java.io.ByteArrayOutputStream;
 import java.io.FileDescriptor;
 import java.io.FileOutputStream;
@@ -101,6 +102,7 @@ class WebsocketLauncherTest {
     session = webSocketContainer.connectToServer(client, new URI("ws://localhost:" + websocketPort + endpointPath));
   }
 
+  @SuppressFBWarnings("RV_RETURN_VALUE_IGNORED")
   void testWebsocketServer(int websocketPort) {
     try {
       latch.await(1, TimeUnit.SECONDS);
